Bot that mixes between different bots

The first time the bot is called for a particular player in a play He picks a child bot randomly. Then it continues with that child bot for this player the whole play.

bot_mixture(game, player, child_bots, prob = NULL, ...)

Arguments

game
the game object

player
the player number of this bot

child_bots
A list of child pots

prob
A vector of weights for each child bot. If NULL (default) all are equally likely.

Details

If you use _bots or call repeatedly bot_mixture to generate mixture bots for each player, the child bots will be independently drawn for each player.

Instead, if bot1 is a mixture bot for player 1 and you create a mixture bot for player 2 by bot2 = bot12$player = 2 then bot2 will select in every play the same child bot than bot1.

See also